mpprint: %f (et al) will never be used in mp_printf
.. this does not affect Python `"%f" % 3.14` which uses
mp_print_float directly.

Saves 40 bytes in trinket_m0.

+++ ports/unix/coverage.c
@@ -447,9 +447,9 @@ STATIC mp_obj_t extra_coverage(void) {
         float far[1];
         double dar[1];
         mp_binary_set_val_array_from_int('f', far, 0, 123);
-        mp_printf(&mp_plat_print, "%.0f\n", (double)far[0]);
+        mp_printf(&mp_plat_print, "%d\n", (int)far[0]);
         mp_binary_set_val_array_from_int('d', dar, 0, 456);
-        mp_printf(&mp_plat_print, "%.0lf\n", dar[0]);
+        mp_printf(&mp_plat_print, "%d\n", (int)dar[0]);
     }
 
     // VM
